[Fix] MASTER는 모임탈퇴할 수 없도록 exception처리

+++ src/main/java/stg/onyou/service/ClubService.java
@@ -1040,6 +1040,10 @@ public class ClubService {
             throw new CustomException(ErrorCode.NO_PERMISSION);
         }
 
+        if( userClub.getRole().equals(Role.MASTER) ){
+            throw new CustomException(ErrorCode.MASTER_WITHDRAW_EXCEPTION);
+        }
+
         userClubRepository.delete(userClub);
 
     }
